# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "DocumentStatus": {
        "description": "Upload-progress polling: where a freshly ingested document is.",
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/PipelineState"
          }
        },
        "required": [
          "id",
          "state",
          "label",
          "error"
        ],
        "title": "DocumentStatus",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Body of every non-2xx ``/api/v1`` response.\n\n``code`` is a stable machine-readable identifier the client branches on;\n``detail`` is the human-readable message it may show as-is.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "PipelineState": {
        "enum": [
          "pending",
          "running",
          "completed",
          "failed",
          "dismissed",
          "partial"
        ],
        "title": "PipelineState",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "Sanctuary API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/documents/{doc_id}": {
      "delete": {
        "operationId": "v1_delete_document_api_v1_documents__doc_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "V1 Delete Document",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/api/v1/documents/{doc_id}/status": {
      "get": {
        "description": "Upload-progress polling: terminal states carry no further changes.",
        "operationId": "v1_document_status_api_v1_documents__doc_id__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "V1 Document Status",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/api/v1/triage/documents/{doc_id}": {
      "delete": {
        "operationId": "v1_delete_loose_api_v1_triage_documents__doc_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "V1 Delete Loose",
        "tags": [
          "triage"
        ]
      }
    },
...
  }
}